libs/racing-coach-core/pyproject.toml
[project]
name = "racing-coach-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Sawyer", email = "sawyerczupka@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "maturin>=1.10.2",
    "numpy>=1.26.0",
    "pandas>=2.2.3",
    "pandas-stubs>=2.3.3.251201",
    "plotly>=6.3.0",
    "pyarrow>=15.0.0",
    "pyarrow-stubs>=20.0.0.20251209",
    "pydantic>=2.11.2",
    "pyirsdk>=1.3.5",
    "racing-coach-server-client",
]

[tool.uv.sources]
racing-coach-server-client = { path = "../racing-coach-api-client", editable = true }

[build-system]
requires = ["maturin>=1.7,<2.0"]
build-backend = "maturin"

[tool.maturin]
# Build a mixed Rust/Python project (src layout)
# See: https://www.maturin.rs/project_layout.html
python-source = "src"
manifest-path = "rust/Cargo.toml"
module-name = "racing_coach_core._rs"
features = ["pyo3/extension-module"]

# [tool.isort]
# known_first_party = ["racing_coach_core"]

[tool.basedpyright]
typeCheckingMode = "strict"
pythonVersion = "3.13"
enableTypeIgnoreComments = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "unit: Unit tests that mock external dependencies",
    "integration: Integration tests that test components together",
    "slow: Tests that take a long time to run",
    "load: Load/stress tests for high-frequency event processing (skip in CI)",
]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=racing_coach_core.events",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[dependency-groups]
dev = [
    "openapi-python-client>=0.27.1",
    "ruff>=0.12.11",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "factory-boy>=3.3.0",
    "pytest-factoryboy>=2.8.0",
    "types-factory-boy>=0.4.1",
    "pytest-randomly>=4.0.1",
]
